+++ src/arrayIO/vector.h
@@ -16,7 +16,7 @@ template <class T> class VecNumWriter final : public BulkNumWriter<T> {
   public:
     VecNumWriter(std::vector<T> &vec) : vec(vec) {}
 
-    uint32_t write(T *in, uint32_t count) override {
+    uint64_t write(T *in, uint64_t count) override {
         size_t initial_size = vec.size();
         vec.resize(vec.size() + count);
         std::memmove(vec.data() + initial_size, in, sizeof(T) * count);
@@ -29,21 +29,21 @@ using VecUIntWriter = VecNumWriter<uint32_t>;
 template <class T> class VecNumReader : public BulkNumReader<T> {
   private:
     const T *vec;
-    uint32_t capacity;
-    uint32_t pos = 0;
+    uint64_t capacity;
+    uint64_t pos = 0;
 
   public:
     VecNumReader(const T *vec, std::size_t capacity) : vec(vec), capacity(capacity) {}
 
     // Return total number of integers in the reader
-    uint32_t size() const override { return capacity; }
+    uint64_t size() const override { return capacity; }
 
     // Change the next load to start at index pos
-    void seek(uint32_t new_pos) override { pos = new_pos; }
+    void seek(uint64_t new_pos) override { pos = new_pos; }
 
     // Copy up to `count` integers into `out`, returning the actual number copied.
     // Will always load >0 unless there is no more input
-    uint32_t load(T *out, uint32_t count) override {
+    uint64_t load(T *out, uint64_t count) override {
         std::memmove(out, vec + pos, sizeof(T) * count);
         pos += count;
         return count;
@@ -69,10 +69,10 @@ class VecReaderWriterBuilder : public WriterBuilder, public ReaderBuilder {
     std::map<std::string, std::vector<double>> double_vecs;
     std::map<std::string, std::vector<std::string>> string_vecs;
     std::string version;
-    uint32_t chunk_size;
+    uint64_t chunk_size;
 
   public:
-    VecReaderWriterBuilder(uint32_t chunk_size = 1024);
+    VecReaderWriterBuilder(uint64_t chunk_size = 1024);
     UIntWriter createUIntWriter(std::string name) override;
     ULongWriter createULongWriter(std::string name) override;
     FloatWriter createFloatWriter(std::string name) override;
